backend/requirements.txt
# ============================================
# ConvergeAI Backend Dependencies
# ============================================
# Python 3.11+ required

# ============================================
# Core Framework
# ============================================
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
python-dotenv==1.0.1

# ============================================
# Database - MySQL (Async)
# ============================================
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
aiomysql==0.2.0
pymysql==1.1.1
cryptography==44.0.0

# ============================================
# Redis (Async)
# ============================================
redis[hiredis]==5.2.1
aioredis==2.0.1

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pyjwt==2.10.1

# ============================================
# LangChain & LangGraph
# ============================================
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-google-genai==2.0.8
langgraph==0.2.59
langsmith==0.2.7

# ============================================
# Google Generative AI (Gemini)
# ============================================
google-generativeai==0.8.3
google-cloud-aiplatform==1.75.0

# ============================================
# Vector Database - Pinecone
# ============================================
pinecone-client==5.0.1
pinecone[grpc]==5.0.1

# ============================================
# NLP & ML Libraries
# ============================================
transformers==4.47.1
torch==2.5.1
sentence-transformers==3.3.1
numpy==1.26.4
scikit-learn==1.6.0

# ============================================
# Task Queue - Celery
# ============================================
celery[redis]==5.4.0
flower==2.0.1

# ============================================
# HTTP Client
# ============================================
httpx==0.28.1
aiohttp==3.11.11
requests==2.32.3

# ============================================
# Data Validation & Serialization
# ============================================
email-validator==2.2.0
phonenumbers==8.13.51
python-dateutil==2.9.0.post0

# ============================================
# Monitoring & Observability
# ============================================
prometheus-client==0.21.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0

# ============================================
# Logging
# ============================================
python-json-logger==3.2.1
loguru==0.7.3

# ============================================
# Testing
# ============================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.28.1
faker==33.1.0

# ============================================
# Code Quality
# ============================================
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.14.0
pylint==3.3.2
pre-commit==4.0.1

# ============================================
# Documentation
# ============================================
mkdocs==1.6.1
mkdocs-material==9.5.50

# ============================================
# Utilities
# ============================================
python-slugify==8.0.4
pytz==2024.2
tenacity==9.0.0
pyyaml==6.0.2
orjson==3.10.12
ujson==5.10.0

# ============================================
# File Processing
# ============================================
python-magic==0.4.27
pillow==11.0.0
pypdf2==3.0.1
python-docx==1.1.2

# ============================================
# Rate Limiting
# ============================================
slowapi==0.1.9

# ============================================
# Email & Communication
# ============================================
resend==2.6.0
emails==0.6

# ============================================
# CORS
# ============================================
# Included in FastAPI

# ============================================
# Environment-specific
# ============================================
# Development
ipython==8.31.0
ipdb==0.13.13

# Production
gunicorn==23.0.0

